feat: 예약 개념 삭제(Book)
- 기존 동시에 같은 좌석을 예매하지 못하게 만들기 위해서 예약이란 개념을 만들었다.
- 하지만, 결제하는 시간에 해당 좌석을 DB에서 잠금처리 하면 되기 때문에 필요 없는 개념이다.
- 또한, 예약 엔티티를 생성하기 위해서도 DB 잠금을 어차피 해야하기 때문에 필요없다.

+++ src/main/kotlin/com/haphollys/booook/model/PriceList.kt
@@ -5,14 +5,15 @@ import com.haphollys.booook.domains.room.RoomEntity.RoomType.TWO_D
 import com.haphollys.booook.domains.screen.Seat
 import com.haphollys.booook.domains.screen.Seat.SeatType.*
 import org.springframework.stereotype.Component
+import java.math.BigDecimal
 
 @Component
 class PriceList(
-    val table: Map<RoomEntity.RoomType, Map<Seat.SeatType, Int>> = mapOf(
+    val table: Map<RoomEntity.RoomType, Map<Seat.SeatType, BigDecimal>> = mapOf(
         TWO_D to mapOf(
-            FRONT to 1000,
-            MIDDLE to 2000,
-            BACK to 3000
+            FRONT to BigDecimal(1000),
+            MIDDLE to BigDecimal(2000),
+            BACK to BigDecimal(3000)
         ),
     )
 )
