+++ src/main/kotlin/com/haphollys/booook/service/external/pg/PGDto.kt
@@ -1,9 +1,11 @@
 package com.haphollys.booook.service.external.pg
 
+import java.math.BigDecimal
+
 class PGDto {
     data class PaymentRequest(
         val paymentId: Long,
-        val amount: Int
+        val amount: BigDecimal
     )
 
     data class UnPaymentRequest(
